+++ pyodesys/__init__.py
@@ -10,5 +10,4 @@ solvers such as ODEPACK in SciPy, CVode in sundials, GSL or odeint together.
 from __future__ import absolute_import
 
 from .core import OdeSys
-from .symbolic import SymbolicSys
-assert OdeSys, SymbolicSys  # silence pyflakes
+